requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.7
django-ckeditor==6.7.2
django-filter==25.1
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==3.1.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tablib==3.8.0
